clinical_triage_3d/server/requirements.txt
openenv-core>=0.2.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
openai>=1.30.0
httpx>=0.27.0
requests>=2.31.0
pygame>=2.5.0
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
Pillow>=10.0.0
numpy>=1.26.0
